Changes after class today
Changes to:
Ethernet.c - changed names and fixed functions. Changed to 1 set of
functions operating on a structure indexed in an array.
Current thought is index 0 - receive, index 1 - send

ETH_Receive.c
Corrected the state cases for Attached, New packet, and S2.
Not positive this is the correct operation to make however it is the
code logic that was there.

IP_Receive.c
minor fixes.

ETH_Send.h
fixed function declaration error.